moved custom azure form to a separate component
